Refactor FfiCall a bit
